User post (#50)
* Seeds1.0 (#36)

* DB Creation, Table Declaration

* ASSOCIATION: Working User linking to Wishlisted_Strain

* added all associations

* adding in routes

* routes WIP

* routes WIP

* fixed spacing and added use strict to models

* routes WIP

* wip

* routes WIP

* routes WIP'

* all get routes working

* WIP

* WIP

* created all gibberish data rows for db

* fixed a few seed entries

* added merchant post routes to add new merchants (#37)

* added routes for newly signed up users and growers

* added routes for newly signed up users and growers (#38)

* modified grower model

* Working on adding strain review and updating the avg ratings of strain and label

* update routes working

* WIP update avgRating when new review is added

* modified rating column, update avg_rating working for strains and labels

+++ models/Label.js
@@ -17,7 +17,7 @@ module.exports = function (sequelize, DataTypes) {
       allowNull: false
     },
     label_avg_rating: {
-      type: DataTypes.INTEGER
+      type: DataTypes.DECIMAL(10,2)
     }
   })
   Label.associate = models => {
